+++ packages/make-dedicated-lockfile/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @pnpm/make-dedicated-lockfile
 
+## 0.2.22
+
+### Patch Changes
+
+- Updated dependencies [e1b459008]
+  - @pnpm/exportable-manifest@2.3.0
+
 ## 0.2.21
 
 ### Patch Changes